refactor(auth): move AuthUserDropdown to auth ui module
